fix: add static_cast<DWORD> for WaitForSingleObject timeout on MSVC
std::chrono::milliseconds::rep is int64 on MSVC, but WaitForSingleObject
expects DWORD (uint32). MSVC warning C4244 with /WX becomes an error.
Add explicit static_cast to suppress the narrowing warning.

+++ include/subprocess/subprocess.hpp
@@ -2434,7 +2434,8 @@ class subprocess {
     }
     auto rc = WaitForSingleObject(
         process_handle_.get(),
-        timeout_.has_value() ? timeout_.value().count() : INFINITE);
+        timeout_.has_value() ? static_cast<DWORD>(timeout_.value().count())
+                             : INFINITE);
     if (rc == WAIT_OBJECT_0) {
       DWORD ret{127};
       GetExitCodeProcess(process_handle_.get(), &ret);
